feat(pricing): mirror /pilot to /pricing, extract shared PilotApplication
Pricing is pilot-gated (per §25 company-context.md — "be clear it's
pilot-gated"). Having /pricing dead-end on a ComingSoon card while
/pilot has the real application form was the worst of both worlds:
users clicking Pricing bounced out, and we lost every qualified
applicant coming through pricing intent.

- Extract pilot page content into components/sections/PilotApplication
  (shared JSX: eyebrow, H1, lede, two context cards, form).
- /pilot/page.tsx now a thin wrapper (header + PilotApplication + footer)
  with pilot-specific metadata.
- /pricing/page.tsx: replace ComingSoon placeholder with PilotApplication,
  distinct metadata ("Pricing · Salency") so SEO separates the URLs.

Both routes render identical content. Trivial to diverge later when
public pricing terms exist.
